Broker commitlog (#80)
* Detect replaced connections in router

* Add link

* Initial acceptor implementation

* Send connack

* Make identifier intrinsic to log

* Update journal

* Add mesh and tracker

* Implement parts of mesh codec

* More implementation

* Implement encoder

* Move LinkHandle

* Send u64 as packet ids and send 1 Data for replication

* Represent connections with usize instead of string

* Distinguish replicators and device connection in router and add connection ack support

* Add pkid to Data

* Add DataAck

* Separate segemnt id, offset and count for native and replicated data

* Consider data from replicated log while extracting data

* Update journal

* Separate replicated data and native data extraction

* Extract topics request into a new method

* Reply with empty response when extraction fails

* More focussed extraction

* Fix unnecessary increment in offset while registering data notification

* Test script to track throughput every commit

* Register writer connection in example to receive acks

* Use async channel for some free gains

* Update watermarks method

* Reorganize methods

* Add watermark triggers

* Use single reply method

* Register only connection ids for watermark notifications

* Add more methods

* Reply topic requests with empty responses

* Reply topic requests with empty responses
Make topics extraction code consistent with data extraction

* Implement watermark empty responses

* Add topic count to example

* Add topic and watermark tracker

+++ rumqttlog/src/volatile/index.rs
@@ -5,7 +5,7 @@ pub struct Index {
     /// Base offset
     base_offset: u64,
     /// Ids of the payloads
-    ids: Vec<u16>,
+    ids: Vec<u64>,
     /// Positions of the payloads
     positions: Vec<u64>,
     /// Sizes of the payloads
@@ -31,14 +31,14 @@ impl Index {
         self.sizes.len() as u64
     }
 
-    pub fn write(&mut self, id: u16, pos: u64, len: u64) {
+    pub fn write(&mut self, id: u64, pos: u64, len: u64) {
         self.ids.push(id);
         self.positions.push(pos);
         self.sizes.push(len);
     }
 
     /// Reads an offset from the index and returns segment record's offset, size and id
-    pub fn read(&self, offset: u64) -> io::Result<(u64, u64, u16)> {
+    pub fn read(&self, offset: u64) -> io::Result<(u64, u64, u64)> {
         if self.positions.len() == 0 {
             return Err(io::Error::new(io::ErrorKind::UnexpectedEof, "No entries in index"));
         }
@@ -61,7 +61,7 @@ impl Index {
     /// Returns starting position, size required to fit 'n' records, n (count)
     /// Total size of records might cross the provided boundary. Use returned size
     /// for allocating the buffer
-    pub fn readv(&self, offset: u64, target_size: u64) -> io::Result<(u64, u64, Vec<u16>)> {
+    pub fn readv(&self, offset: u64, target_size: u64) -> io::Result<(u64, u64, Vec<u64>)> {
         let mut count = 0;
         let mut current_size = 0;
         let offset = offset as usize;
